test: added first few basic mint tests

+++ src/struct/JB721StakingTier.sol
@@ -6,6 +6,6 @@ pragma solidity ^0.8.16;
  * @param amount the amount to stake (has to be more than the tier minimum amount)
  */
 struct JB721StakingTier {
-    uint8 tierId;
+    uint16 tierId;
     uint128 amount;
 }
